Initial commit for file processor
This is an initial commit for the file processor component of the overall solution. It is anticipated that a file will be uploaded to an Azure Storage Account, and the file processor will kick in upon completion of a file being uploaded. This commit validates that the file is not empty, and is a JSON file that has at least one product and a transmission summary with all properties populated.