Refactoring of Functions class
The Functions class has been refactored so that all code that validates whether an incoming JSON file is valid is contained in a new class (FileValidator). Intention is that the Functions class will act more as an orchestrator, passing off responsibilities of validating the incoming file and recording data in the database to injected dependencies.