Deploy infrastructure stack using nested ARM templates
The existing ARM templates needed some updates to provide output that is then used for provisioning another resource. For example, once the SQL Server container instance is deployed, return the connection string so that it can then be specified as an application setting (environment variable) when provisioning the container instance for the File Processor application.