ARM template outputs for SecureString changed to String
When deploying the stack using nested ARM templates, the issue was that sensitive output such as storage account connection string and database connection string were passed from the nested templates to the main template as SecureString. This would then cause the deployment to fail, as the returned SecureString object had no `value` property. This has been resolved by changing these outputs to String. Note that the linked template that provisions the container instance for the File Processor application has such parameters (inputs) defined as SecureString, meaning that these values cannot be obtained from deployment logs for that component. Unfortunately, they can be obtained from the outputs of the deployment logs for the storage account and SQL Server container instance.