Attempt to mount volume to DB container
ARM templates have been updated to provision a file share in the provisioned Azure Storage Account, and to mount the Storage Account file share to the database Azure Container Instance. The aim of this is for the script that deploys the database to do so in the mounted file share, rather than directly on the container. That way, if a new container instance is provisioned, the database is not lost.