Check that login and user do not already exist
When provisioning the database in the running Azure Container Instance, it may be possible that the Login and User to be used by the File Processor Azure WebJob may already exist. Thus, only create the Login if it does not already exist, and only create the User if it does not already exist in the warehouse database.