Attempt to create database in Storage Account file share
For the databse to be more permanent, an attempt has been made for the database to be created in the Storage Account file share that has been mounted to the container running SQL Server. Currently the database is created in the running container, which means that if the container is deleted, so too is the database.